feat: Update environment variable handling logic
Updated .env file handling logic into a separate file to improve
separation of concerns and declutter the main file. Updated the
makefile with a revised "run" command and the dev.Dockerfile to
reflect this change. The dev.Dockerfile was also updated to
pre-install gcc and libc-dev packages for debugging purposes.